+++ source/B2B.Transactions.CimMessageAdapter.Tests/Stubs/MessageIdsStub.cs
@@ -15,7 +15,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using B2B.CimMessageAdapter;
-using B2B.CimMessageAdapter.Message.MessageIds;
+using B2B.CimMessageAdapter.Messages;
 
 namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests.Stubs
 {
