+++ source/B2B.Transactions.CimMessageAdapter/B2B.Transactions.CimMessageAdapter.csproj
@@ -29,4 +29,8 @@ limitations under the License.
     <ItemGroup>
       <ProjectReference Include="..\Energinet.DataHub.MarketRoles.Infrastructure\Energinet.DataHub.MarketRoles.Infrastructure.csproj" />
     </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Message\MessageIds" />
+    </ItemGroup>
 </Project>
